GraphQLクエリを利用するcomponentに近い位置に移した
